requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
pillow
psycopg2-binary
drf-yasg
python-decouple
celery==5.2.7
django-rest-passwordreset==1.2.1